Added better error handling for fao_areas
Seems you can trigger a "database unavailable" error by making too many
requests on the HTML.

I've added a trigger to pause and try again if the database doesn't
respond, which should hopefully avoid returning a NULL when the data is
in fact on fishbase.  If the second attempt fails, the function will
return a NULL for that species and should continue on to the next one
rather than throwing an error.